requirements.txt
tensorflow-gpu==2.7
matplotlib
PIL
numpy